backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic-settings==2.5.0
faster-whisper==1.0.3
piper-tts==1.2.0
httpx==0.27.0
python-multipart==0.0.9
